+++ src/frontend/components/styles/OptionDetailsView.module.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "src/styles/variables.scss";
 
 
 @keyframes fadeIn {
