+++ src/frontend/components/styles/OptionRequestForm.module.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "src/styles/variables.scss";
 
 @keyframes fadeIn {
     0% { opacity: 0; }
