+++ src/frontend/components/styles/ThemeSwitch.module.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "src/styles/variables.scss";
 
 button.button {
     display: inline;
