+++ src/pages/styles/404.module.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "src/styles/variables.scss";
 
 .root {
     display: flex;
