+++ src/pages/styles/LandingPage.module.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "src/styles/variables.scss";
 
 .root {
     max-width: 80%;
