+++ src/pages/styles/Team.module.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "src/styles/variables.scss";
 
 .root {
     margin: auto;
