+++ src/pages/styles/_app.scss
@@ -1,5 +1,5 @@
 @import url("https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap");
-@import "../variables.scss";
+@import "src/styles/variables.scss";
 
 html {
     min-height: 100%;
